Rename system_simulation file to system_component
