+++ docs/user/reference/api.rst
@@ -56,10 +56,10 @@ This is the internal API reference for tickit
             ``tickit.core.components.device_component``
             --------------------------------------------
 
-        .. automodule:: tickit.core.components.system_simulation
+        .. automodule:: tickit.core.components.system_component
             :members:
 
-            ``tickit.core.components.system_simulation``
+            ``tickit.core.components.system_component``
             --------------------------------------------
