+++ examples/configs/nested.yaml
@@ -2,7 +2,7 @@
   name: random_trampoline
   inputs: {}
   callback_period: 10000000000
-- type: tickit.core.components.system_simulation.SystemSimulation
+- type: tickit.core.components.system_component.SystemSimulation
   name: internal_tickit
   inputs:
     input_1:
