+++ src/tickit/core/components/system_component.py
