+++ tests/core/components/test_system_simulation.py
@@ -7,7 +7,7 @@ from mock import AsyncMock, Mock, patch
 from mock.mock import create_autospec
 
 from tickit.core.components.component import Component
-from tickit.core.components.system_simulation import (
+from tickit.core.components.system_component import (
     SystemSimulation,
     SystemComponent,
 )
@@ -39,7 +39,7 @@ def mock_state_consumer_type() -> Mock:
 
 @pytest.fixture
 def patch_scheduler() -> Iterable[Mock]:
-    spec = "tickit.core.components.system_simulation.SlaveScheduler"
+    spec = "tickit.core.components.system_component.SlaveScheduler"
     with patch(spec, autospec=True) as mock:
 
         def on_tick(time, changes):
@@ -72,7 +72,7 @@ def test_system_simulation_constructor(system_simulation: SystemSimulation):
 @pytest.fixture
 def patch_asyncio() -> Iterable[Mock]:
     with patch(
-        "tickit.core.components.system_simulation.asyncio", autospec=True
+        "tickit.core.components.system_component.asyncio", autospec=True
     ) as mock:
         yield mock
